Extract Game Properties
