+++ GameProperties.py
@@ -1,0 +1,10 @@
+class GameProperties:
+    white = (255, 255, 255)
+    black = (0, 0, 0)
+    red = (255, 0, 0)
+    green = (0, 255, 0)
+    blue = (0, 0, 255)
+    width = 1200
+    height = 600
+    snake_blok = 10
+    snake_speed = 15
