+++ zadanie.py
@@ -1,19 +1,8 @@
 import pygame
 import random
+from GameProperties import *
 
 pygame.init()
-
-
-class GameProperties:
-    white = (255, 255, 255)
-    black = (0, 0, 0)
-    red = (255, 0, 0)
-    green = (0, 255, 0)
-    blue = (0, 0, 255)
-    width = 1200
-    height = 600
-    snake_blok = 10
-    snake_speed = 15
 
 
 screen = pygame.display.set_mode((GameProperties.width, GameProperties.height))
